Refactor the root class Timers into a module for easier composition and additional functionality going forward.
